split WSL-backed local init
